+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ### Patch Changes
 
+- Updated dependencies [29264a3]
+  - ai@6.0.3
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [129ff26]
   - ai@6.0.2